// frontend/src/lib/types.ts
export interface ReplaceTextOp {
  op: "replaceText";
  selector: string;
  new_text: string;
  original_text: string | null;
}

export interface ReplaceAttributeOp {
  op: "replaceAttribute";
  selector: string;
  attribute: string;
  new_value: string;
}

export interface AddClassOp {
  op: "addClass";
  selector: string;
  class_name: string;
}

export interface RemoveClassOp {
  op: "removeClass";
  selector: string;
  class_name: string;
}

export interface InsertBeforeOp {
  op: "insertBefore";
  selector: string;
  html: string;
}

export interface InsertAfterOp {
  op: "insertAfter";
  selector: string;
  html: string;
}

export interface ReplaceStyleOp {
  op: "replaceStyle";
  selector: string;
  css_text: string;
}

export type PatchOperation =
  | ReplaceTextOp
  | ReplaceAttributeOp
  | AddClassOp
  | RemoveClassOp
  | InsertBeforeOp
  | InsertAfterOp
  | ReplaceStyleOp;

export interface PatchSpec {
  variant_id: string;
  description: string;
  operations: PatchOperation[];
}